requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
python-telegram-bot>=21.0
anthropic>=0.25.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
